Moving service_client.hpp to nav2_util
The ServiceClient class will be used Lifecycle helper functions that
will be put in nav2_util, so we need it there. It seemed more of utility
class than a Task anyhow.
