+++ nav2_tasks/include/nav2_tasks/costmap_service_client.hpp
@@ -15,24 +15,24 @@
 #ifndef NAV2_TASKS__COSTMAP_SERVICE_CLIENT_HPP_
 #define NAV2_TASKS__COSTMAP_SERVICE_CLIENT_HPP_
 
-#include "nav2_tasks/service_client.hpp"
+#include "nav2_util/service_client.hpp"
 #include "nav2_msgs/srv/get_costmap.hpp"
 
 namespace nav2_tasks
 {
 
-class CostmapServiceClient : public ServiceClient<nav2_msgs::srv::GetCostmap>
+class CostmapServiceClient : public nav2_util::ServiceClient<nav2_msgs::srv::GetCostmap>
 {
 public:
   CostmapServiceClient()
-  : ServiceClient<nav2_msgs::srv::GetCostmap>("GetCostmap")
+  : nav2_util::ServiceClient<nav2_msgs::srv::GetCostmap>("GetCostmap")
   {
   }
 
   using CostmapServiceRequest =
-    ServiceClient<nav2_msgs::srv::GetCostmap>::RequestType;
+    nav2_util::ServiceClient<nav2_msgs::srv::GetCostmap>::RequestType;
   using CostmapServiceResponse =
-    ServiceClient<nav2_msgs::srv::GetCostmap>::ResponseType;
+    nav2_util::ServiceClient<nav2_msgs::srv::GetCostmap>::ResponseType;
 };
 
 }  // namespace nav2_tasks
